Added weapons and base FSM

+++ Assets/Scripts/IDamagable.cs
@@ -1,5 +1,6 @@
+
 public interface IDamagable
 {
-    int Health { get; }
-    void TakeDamage(int damage);
+    float Health { get; }
+    void TakeDamage(float damage);
 }
